Make attributes available in new planner interface
* Protos.Attribute is translated to planner.spi.Attribute
* planner.spi.Offer includes both attributes and resources
* A few more to add to planner.spi.Offer
* Change protocol.data.Range from (int, int) to (long, long)
* It's a bit strange attributes in offers are translated to POJO
  while attributes in protocol.data.Job is just a String

+++ retz-common/src/test/java/io/github/retz/protocol/RangeTest.java
@@ -33,7 +33,7 @@ public class RangeTest {
 
         r = Range.parseRange("2-");
         assertEquals(2, r.getMin());
-        assertEquals(Integer.MAX_VALUE, r.getMax());
+        assertEquals(Long.MAX_VALUE, r.getMax());
 
         int[] list = {2, 3, 4, 5, 10, 1024};
         for (int i : list) {
@@ -44,11 +44,11 @@ public class RangeTest {
 
         r = Range.parseRange("0-0");
         assertEquals(0, r.getMin());
-        assertEquals(Integer.MAX_VALUE, r.getMax());
+        assertEquals(Long.MAX_VALUE, r.getMax());
 
         r = Range.parseRange("2-0");
         assertEquals(2, r.getMin());
-        assertEquals(Integer.MAX_VALUE, r.getMax());
+        assertEquals(Long.MAX_VALUE, r.getMax());
     }
 
     @Test(expected = IllegalArgumentException.class)
